+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Make Inventory search and browse query boxes expandable. Refs UIIN-2493.
 * Added support for `containsAny` match option in Advanced search. Refs UIIN-2486.
 * Inventory search/browse: Do not retain checkbox selections when toggling search segment. Refs UIIN-2477.
+* Search box/Browse box- Reset all should shift focus back to search box. Refs UIIN-2514.
 
 ## [10.0.3] IN PROGRESS